modularize state map events
